requirements.txt
torch>=2.1.0
numpy>=1.26.0
pandas>=2.1.0
scikit-learn>=1.3.2
matplotlib>=3.8.0
seaborn>=0.13.0
pyyaml>=6.0.1
tqdm>=4.66.0
tensorboard>=2.15.0
pyarrow>=14.0.0
mamba-ssm>=1.2.0
modular>=0.5.0
pandas-ta>=0.3.14b
arch>=6.2.0
joblib>=1.3.0
rich>=13.7.0
